Finish all the exercises except algorithm part

+++ exercises/functions/functions3.rs
@@ -3,13 +3,12 @@
 // Execute `rustlings hint functions3` or use the `hint` watch subcommand for a
 // hint.
 
-
 fn main() {
     call_me(3);
 }
 
 fn call_me(num: u32) {
-    for i in 0..num {
+    for i in 0..num as usize {
         println!("Ring! Call number {}", i + 1);
     }
 }
